+++ web/app/pages/styleGuide/TabsPage.js
@@ -3,33 +3,24 @@ import React from 'react';
 
 import CodeExample from '../../react/components/CodeExample';
 
-import InteractiveExample from './assets/examples/Tabs/TabsInteractiveExample';
-
-import Tab from 'src/components/Tabs/Tabs';
-import TabItem from 'src/components/Tabs/Tab';
+import tabsBasicExample from 'raw-loader!app/pages/styleGuide/assets/examples/Tabs/TabsBasic.example';
 
 class TabsPage extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {};
-  }
+  static contextTypes = {
+    router: PropTypes.object.isRequired,
+  };
 
   render() {
     return (
       <div className="p-style-guide-tabs">
         <h1>Tabs</h1>
         <CodeExample
-          exampleComponent={InteractiveExample}
-          codeContent=""
+          axis="horizontal"
+          example={tabsBasicExample}
         />
       </div>
     );
   }
 }
-
-TabsPage.contextTypes = {
-  router: PropTypes.object.isRequired
-};
 
 export default TabsPage;