+++ web/app/pages/styleGuide/assets/examples/Buttons/ButtonBasic.example.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+class Example extends PureComponent {
+  render() {
+    return <Button>test</Button>;
+  }
+}
+
+ReactDOM.render(<Example />, mountNode);