+++ web/app/pages/styleGuide/assets/examples/SvgIcons/SvgIconBasic.example.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+class Example extends PureComponent {
+  render() {
+    return <SvgIcon icon="Person" />;
+  }
+}
+
+ReactDOM.render(<Example />, mountNode);